expand dev-files architecture
